@@ -1,11 +1,18 @@
 ﻿// Copyright (c) Matteo Ciapparelli.
 // Licensed under the MIT license.
 
-using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Options;
 using System.Diagnostics.CodeAnalysis;
+
+#if NET9_0_OR_GREATER
+using Microsoft.Extensions.Caching.Hybrid;
+#else
+
+using Microsoft.Extensions.Caching.Distributed;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+
+#endif
 
 namespace FeatureManagement.Database;
 
@@ -15,13 +22,18 @@ namespace FeatureManagement.Database;
 public class CachedFeatureStore : IFeatureStore
 {
     private readonly IFeatureStore _featureStore;
+#if NET9_0_OR_GREATER
+    private readonly HybridCache _cache;
+#else
     private readonly IDistributedCache _cache;
-    private readonly FeatureCacheOptions _cacheOptions;
 
-    private readonly JsonSerializerOptions _jsonOptions = new JsonSerializerOptions(JsonSerializerOptions.Default)
+    private readonly JsonSerializerOptions _jsonOptions = new(JsonSerializerOptions.Default)
     {
         ReferenceHandler = ReferenceHandler.IgnoreCycles
     };
+
+#endif
+    private readonly FeatureCacheOptions _cacheOptions;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="CachedFeatureStore"/> class.
@@ -29,7 +41,12 @@ public class CachedFeatureStore : IFeatureStore
     /// <param name="featureStore">The concrete feature store.</param>
     /// <param name="cache">The cache service.</param>
     /// <param name="options">The cache options.</param>
-    public CachedFeatureStore(IFeatureStore featureStore, IDistributedCache cache,
+    public CachedFeatureStore(IFeatureStore featureStore,
+#if NET9_0_OR_GREATER
+        HybridCache cache,
+#else
+        IDistributedCache cache,
+#endif
         IOptions<FeatureCacheOptions> options)
     {
         _featureStore = featureStore ?? throw new ArgumentNullException(nameof(featureStore));
@@ -69,23 +86,56 @@ public class CachedFeatureStore : IFeatureStore
 
     private async Task<TData> GetCacheAsync<TData>(string key) where TData : class
     {
-        var cachedData = await _cache.GetAsync(FeatureCacheOptions.CachePrefix + key);
+        var cacheKey = FeatureCacheOptions.CachePrefix + key;
+
+#if NET9_0_OR_GREATER
+        return await _cache.GetOrCreateAsync(cacheKey, static (_) => ValueTask.FromResult(default(TData)));
+#else
+        var cachedData = await _cache.GetAsync(cacheKey);
 
         if (cachedData is null)
             return null;
 
         return await JsonSerializer.DeserializeAsync<TData>(new MemoryStream(cachedData), _jsonOptions);
+#endif
     }
 
-    private async Task SetCacheAsync<TData>(string key, TData data) where TData : class
+    private Task SetCacheAsync<TData>(string key, TData data) where TData : class
     {
-        await _cache.SetAsync(FeatureCacheOptions.CachePrefix + key,
-            JsonSerializer.SerializeToUtf8Bytes(data, _jsonOptions), new DistributedCacheEntryOptions
-            {
-                AbsoluteExpiration = _cacheOptions.AbsoluteExpiration,
-                AbsoluteExpirationRelativeToNow = _cacheOptions.AbsoluteExpirationRelativeToNow,
-                SlidingExpiration = _cacheOptions.SlidingExpiration
-            });
+        var cacheKey = FeatureCacheOptions.CachePrefix + key;
+
+#if NET9_0_OR_GREATER
+        TimeSpan? expiration = null;
+        if (_cacheOptions.AbsoluteExpirationRelativeToNow.HasValue)
+        {
+            expiration = _cacheOptions.AbsoluteExpirationRelativeToNow.Value;
+        }
+        else if (_cacheOptions.SlidingExpiration.HasValue)
+        {
+            // Note: HybridCache treats 'Expiration' more like an absolute TTL from creation/access.
+            // Using SlidingExpiration here effectively sets an absolute expiry based on this duration.
+            expiration = _cacheOptions.SlidingExpiration.Value;
+        }
+        else if (_cacheOptions.AbsoluteExpiration.HasValue && _cacheOptions.AbsoluteExpiration > DateTimeOffset.UtcNow)
+        {
+            expiration = _cacheOptions.AbsoluteExpiration.Value - DateTimeOffset.UtcNow;
+        }
+
+        var entryOptions = new HybridCacheEntryOptions
+        {
+            Expiration = expiration > TimeSpan.Zero ? expiration : null
+        };
+
+        return _cache.SetAsync(cacheKey, data, entryOptions).AsTask();
+#else
+        var options = new DistributedCacheEntryOptions
+        {
+            AbsoluteExpiration = _cacheOptions.AbsoluteExpiration,
+            AbsoluteExpirationRelativeToNow = _cacheOptions.AbsoluteExpirationRelativeToNow,
+            SlidingExpiration = _cacheOptions.SlidingExpiration
+        };
+        return _cache.SetAsync(cacheKey, JsonSerializer.SerializeToUtf8Bytes(data, _jsonOptions), options);
+#endif
     }
 
     #endregion Private
