@@ -1,11 +1,18 @@
 // Copyright (c) Matteo Ciapparelli.
 // Licensed under the MIT license.
 
-using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using System.Text.Json;
 using static FeatureManagement.Database.Abstractions.Features;
+
+#if NET9_0_OR_GREATER
+using Microsoft.Extensions.Caching.Hybrid;
+#else
+
+using Microsoft.Extensions.Caching.Distributed;
+using System.Text.Json;
+
+#endif
 
 namespace FeatureManagement.Database.Abstractions.Tests;
 
@@ -22,11 +29,19 @@ public class CachedFeatureStoreTests
         var serviceProvider = serviceCollection.BuildServiceProvider();
 
         var cachedFeatureStore = serviceProvider.GetRequiredService<IFeatureStore>();
+#if NET9_0_OR_GREATER
+        var cache = serviceProvider.GetRequiredService<HybridCache>();
+#else
         var cache = serviceProvider.GetRequiredService<IDistributedCache>();
+#endif
 
         // Act
         var feature = await cachedFeatureStore.GetFeatureAsync(FirstFeature);
+#if NET9_0_OR_GREATER
+        var cachedFeature = await cache.GetOrCreateAsync(FeatureCacheOptions.CachePrefix + FirstFeature, static (_) => ValueTask.FromResult(default(Feature)));
+#else
         var cachedFeature = JsonSerializer.Deserialize<Feature>(await cache.GetAsync(FeatureCacheOptions.CachePrefix + FirstFeature));
+#endif
 
         // Assert
         Assert.True(feature is not null);
@@ -35,7 +50,7 @@ public class CachedFeatureStoreTests
         Assert.Equal(FirstFeature, feature.Name);
         Assert.Equal(feature.Name, cachedFeature.Name);
 
-        Assert.True(feature.Settings.Any());
+        Assert.True(feature.Settings.Count > 0);
         Assert.Equivalent(feature.Settings, cachedFeature.Settings);
 
         Assert.Equal(FeatureFilterType.TimeWindow, feature.Settings.First().FilterType);
@@ -53,12 +68,20 @@ public class CachedFeatureStoreTests
         var serviceProvider = serviceCollection.BuildServiceProvider();
 
         var cachedFeatureStore = serviceProvider.GetRequiredService<IFeatureStore>();
+#if NET9_0_OR_GREATER
+        var cache = serviceProvider.GetRequiredService<HybridCache>();
+#else
         var cache = serviceProvider.GetRequiredService<IDistributedCache>();
+#endif
         var cacheOptions = serviceProvider.GetRequiredService<IOptions<FeatureCacheOptions>>().Value;
 
         // Act
         var features = await cachedFeatureStore.GetFeaturesAsync();
+#if NET9_0_OR_GREATER
+        var cachedFeatures = await cache.GetOrCreateAsync(FeatureCacheOptions.CachePrefix + cacheOptions.KeyNames.AllFeatures, static (_) => ValueTask.FromResult(default(IReadOnlyCollection<Feature>)));
+#else
         var cachedFeatures = JsonSerializer.Deserialize<IReadOnlyCollection<Feature>>(await cache.GetAsync(FeatureCacheOptions.CachePrefix + cacheOptions.KeyNames.AllFeatures));
+#endif
 
         // Assert
         Assert.True(features is not null);
@@ -79,7 +102,7 @@ public class CachedFeatureStoreTests
 
             Assert.Equal(feature.Name, cachedFeature.Name);
 
-            Assert.True(feature.Settings.Any());
+            Assert.True(feature.Settings.Count > 0);
             Assert.Equivalent(feature.Settings, cachedFeature.Settings);
 
             Assert.Equal(FeatureFilterType.TimeWindow, feature.Settings.First().FilterType);
@@ -98,19 +121,26 @@ public class CachedFeatureStoreTests
         var serviceProvider = serviceCollection.BuildServiceProvider();
 
         var cachedFeatureStore = serviceProvider.GetRequiredService<IFeatureStore>();
+#if NET9_0_OR_GREATER
+        var cache = serviceProvider.GetRequiredService<HybridCache>();
+#else
         var cache = serviceProvider.GetRequiredService<IDistributedCache>();
+#endif
 
         // Act
         var feature = await cachedFeatureStore.GetFeatureAsync(FirstFeature);
         await Task.Delay(1200);
+#if NET9_0_OR_GREATER
+        var featureCache = await cache.GetOrCreateAsync(FirstFeature, static (_) => ValueTask.FromResult(default(Feature)));
+#else
         var featureCache = await cache.GetAsync(FirstFeature);
+#endif
 
         // Assert
         Assert.True(feature is not null);
         Assert.True(featureCache is null);
     }
-    
-    
+
     [Fact]
     public async Task GetAllFeaturesFromCachedStoreWhenFeaturesHaveCircularReference()
     {
@@ -122,11 +152,19 @@ public class CachedFeatureStoreTests
         var serviceProvider = serviceCollection.BuildServiceProvider();
 
         var cachedFeatureStore = serviceProvider.GetRequiredService<IFeatureStore>();
+#if NET9_0_OR_GREATER
+        var cache = serviceProvider.GetRequiredService<HybridCache>();
+#else
         var cache = serviceProvider.GetRequiredService<IDistributedCache>();
+#endif
 
         // Act
         var feature = await cachedFeatureStore.GetFeatureAsync(FirstFeature);
+#if NET9_0_OR_GREATER
+        var cachedFeature = await cache.GetOrCreateAsync(FeatureCacheOptions.CachePrefix + FirstFeature, static (_) => ValueTask.FromResult(default(Feature)));
+#else
         var cachedFeature = JsonSerializer.Deserialize<Feature>(await cache.GetAsync(FeatureCacheOptions.CachePrefix + FirstFeature));
+#endif
 
         // Assert
         Assert.True(feature is not null);
@@ -135,6 +173,6 @@ public class CachedFeatureStoreTests
         Assert.Equal(FirstFeature, feature.Name);
         Assert.Equal(feature.Name, cachedFeature.Name);
 
-        Assert.True(feature.Settings.Any());
+        Assert.True(feature.Settings.Count > 0);
     }
 }
